updated MiMC examples
