+++ .vscode/launch.json
@@ -5,7 +5,7 @@
       "name": "Launch",
       "type": "node",
       "request": "launch",
-      "program": "${workspaceRoot}/bin/examples/mimc/mimc128v3.js",
+      "program": "${workspaceRoot}/bin/examples/mimc/mimc256.js",
       "cwd": "${workspaceRoot}/bin",
       "preLaunchTask": "build",
       "env": {
