+++ bin/examples/mimc/mimc128.js
@@ -3,14 +3,20 @@ Object.defineProperty(exports, "__esModule", { value: true });
 // IMPORTS
 // ================================================================================================
 const assert = require("assert");
+const galois_1 = require("@guildofweavers/galois");
+const air_assembly_1 = require("@guildofweavers/air-assembly");
 const index_1 = require("../../index");
 const utils_1 = require("../../lib/utils");
-const air_assembly_1 = require("@guildofweavers/air-assembly");
+const utils_2 = require("./utils");
+// MIMC PARAMETERS
+// ================================================================================================
+const modulus = 2n ** 128n - 9n * 2n ** 32n + 1n;
+const field = galois_1.createPrimeField(modulus);
+const roundConstants = air_assembly_1.prng.sha256(Buffer.from('4d694d43', 'hex'), 64, field);
+const steps = 2 ** 13;
+const seed = 3n;
 // STARK DEFINITION
 // ================================================================================================
-const steps = 2 ** 13;
-const constantCount = 64;
-const seed = 3n;
 // define security options for the STARK
 const options = {
     hashAlgorithm: 'blake2s256',
@@ -20,40 +26,45 @@ const options = {
     wasm: true
 };
 // create the STARK for MiMC computation
-const mimcStark = index_1.instantiate(Buffer.from(`
-(module
-    (field prime 340282366920938463463374607393113505793)
-    (const $alpha scalar 3)
-    (function $mimcRound
-        (result vector 1)
-        (param $state vector 1) (param $roundKey scalar)
-        (add 
-            (exp (load.param $state) (load.const $alpha))
-            (load.param $roundKey)))
-    (export mimc
-        (registers 1) (constraints 1) (steps ${steps})
-        (static
-            (cycle (prng sha256 0x4d694d43 ${constantCount})))
-        (init
-            (param $seed vector 1)
-            (load.param $seed))
-        (transition
-            (call $mimcRound (load.trace 0) (get (load.static 0) 0)))
-        (evaluation
-            (sub
-                (load.trace 1)
-                (call $mimcRound (load.trace 0) (get (load.static 0) 0))))))`), 'mimc', options, new utils_1.Logger(false));
+const mimcStark = index_1.instantiateScript(Buffer.from(`
+define MiMC over prime field (${modulus}) {
+
+    const alpha: 3;
+    
+    static roundConstant: cycle ${utils_1.inline.vector(roundConstants)};
+
+    secret input startValue: element[1];
+
+    // transition function definition
+    transition 1 register {
+        for each (startValue) {
+            init { yield startValue; }
+
+            for steps [1..${steps - 1}] {
+                yield $r0^3 + roundConstant;
+            }
+        }
+    }
+
+    // transition constraint definition
+    enforce 1 constraint {
+        for all steps {
+            enforce transition($r) = $n;
+        }
+    }
+
+}`), options, new utils_1.Logger(false));
 // TESTING
 // ================================================================================================
 // generate control values
-const controls = runMimc(mimcStark.air.field, steps, constantCount, seed);
+const controls = utils_2.runMimc(mimcStark.air.field, steps, roundConstants, seed);
 // set up inputs and assertions
 const assertions = [
     { step: 0, register: 0, value: controls[0] },
     { step: steps - 1, register: 0, value: controls[steps - 1] } // value at last step is equal to result
 ];
 // prove that the assertions hold if we execute MiMC computation with given inputs
-let proof = mimcStark.prove(assertions, [], [seed]);
+let proof = mimcStark.prove(assertions, [[seed]]);
 console.log('-'.repeat(20));
 // serialize the proof
 let start = Date.now();
@@ -70,16 +81,4 @@ console.log('-'.repeat(20));
 mimcStark.verify(assertions, proof);
 console.log('-'.repeat(20));
 console.log(`STARK security level: ${mimcStark.securityLevel}`);
-// MiMC FUNCTION
-// ================================================================================================
-function runMimc(field, steps, constCount, seed) {
-    // build round constants
-    const roundConstants = air_assembly_1.prng.sha256(Buffer.from('4d694d43', 'hex'), constCount, field);
-    const result = [seed];
-    for (let i = 0; i < steps - 1; i++) {
-        let value = field.add(field.exp(result[i], 3n), roundConstants[i % roundConstants.length]);
-        result.push(value);
-    }
-    return result;
-}
 //# sourceMappingURL=mimc128.js.map
