+++ examples/mimc/mimc128Assembly.ts
@@ -1,17 +1,20 @@
 // IMPORTS
 // ================================================================================================
 import * as assert from 'assert';
-import { StarkOptions, FiniteField } from '@guildofweavers/genstark';
+import { StarkOptions } from '@guildofweavers/genstark';
 import { instantiate } from '../../index';
 import { Logger } from '../../lib/utils';
 import { prng } from '@guildofweavers/air-assembly';
+import { runMimc } from './utils';
 
-// STARK DEFINITION
+// MIMC PARAMETERS
 // ================================================================================================
 const steps = 2**13;
 const constantCount = 64;
 const seed = 3n;
 
+// STARK DEFINITION
+// ================================================================================================
 // define security options for the STARK
 const options: StarkOptions = {
     hashAlgorithm   : 'blake2s256',
@@ -22,49 +25,36 @@ const options: StarkOptions = {
 };
 
 // create the STARK for MiMC computation
-// TODO: update
 const mimcStark = instantiate(Buffer.from(`
 (module
     (field prime 340282366920938463463374607393113505793)
-    (const 
-        (scalar 3))
-    (static
-        (input secret vector (steps ${steps}) (shift -1))
-        (mask inverted (input 0))
-        (cycle (prng sha256 0x4d694d43 ${constantCount})))
-    (transition
-        (span 1) (result vector 1)
-        (local vector 1)
-        (store.local 0 
-			(add 
-				(exp (load.trace 0) (load.const 0))
-                (get (load.static 0) 2)))
-        (add
-            (mul (load.local 0)	(get (load.static 0) 1))
-			(get (load.static 0) 0)
-        )
-    )
-    (evaluation
-        (span 2) (result vector 1)
-		(local vector 1)
-        (store.local 0 
-			(add 
-				(exp (load.trace 0) (load.const 0))
-                (get (load.static 0) 2)))
-        (sub
-            (load.trace 1)
-            (add
-				(mul (load.local 0)	(get (load.static 0) 1))
-				(get (load.static 0) 0))
-		)
-	)
-    (export main (init seed) (steps ${steps})))`
+    (const $alpha scalar 3)
+    (function $mimcRound
+        (result vector 1)
+        (param $state vector 1) (param $roundKey scalar)
+        (add 
+            (exp (load.param $state) (load.const $alpha))
+            (load.param $roundKey)))
+    (export mimc
+        (registers 1) (constraints 1) (steps ${steps})
+        (static
+            (cycle (prng sha256 0x4d694d43 ${constantCount})))
+        (init
+            (param $seed vector 1)
+            (load.param $seed))
+        (transition
+            (call $mimcRound (load.trace 0) (get (load.static 0) 0)))
+        (evaluation
+            (sub
+                (load.trace 1)
+                (call $mimcRound (load.trace 0) (get (load.static 0) 0))))))`
 ), 'mimc', options, new Logger(false));
 
 // TESTING
 // ================================================================================================
 // generate control values
-const controls = runMimc(mimcStark.air.field, steps, constantCount, seed);
+const roundConstants = prng.sha256(Buffer.from('4d694d43', 'hex'), 64, mimcStark.air.field);
+const controls = runMimc(mimcStark.air.field, steps, roundConstants, seed);
 
 // set up inputs and assertions
 const assertions = [
@@ -73,7 +63,7 @@ const assertions = [
 ];
 
 // prove that the assertions hold if we execute MiMC computation with given inputs
-let proof = mimcStark.prove(assertions, [[seed]], [seed]);
+let proof = mimcStark.prove(assertions, [], [seed]);
 console.log('-'.repeat(20));
 
 // serialize the proof
@@ -93,18 +83,3 @@ console.log('-'.repeat(20));
 mimcStark.verify(assertions, proof);
 console.log('-'.repeat(20));
 console.log(`STARK security level: ${mimcStark.securityLevel}`);
-
-// MiMC FUNCTION
-// ================================================================================================
-function runMimc(field: FiniteField, steps: number, constCount: number, seed: bigint): bigint[] {
-    // build round constants
-    const roundConstants = prng.sha256(Buffer.from('4d694d43', 'hex'), constCount, field);
-
-    const result = [seed];
-    for (let i = 0; i < steps - 1; i++) {
-        let value = field.add(field.exp(result[i], 3n), roundConstants[i % roundConstants.length]);
-        result.push(value);
-    }
-
-    return result;
-}
